+++ getaway/interfaces/src/main/java/ar/edu/itba/getaway/interfaces/services/ReviewService.java
@@ -15,9 +15,9 @@ public interface ReviewService {
     Long getReviewAverageScore(ExperienceModel experience);
     List<Long> getListOfAverageScoreByExperienceList(List<ExperienceModel> experienceModelList);
     List<List<Long>> getListOfAverageScoreByExperienceListAndCategoryId(List<List<ExperienceModel>> experienceModelList);
-    Integer getReviewCount(ExperienceModel experience);
-    List<Integer> getListOfReviewCountByExperienceList(List<ExperienceModel> experienceModelList);
-    List<List<Integer>> getListOfReviewCountByExperienceListAndCategoryId(List<List<ExperienceModel>> experienceModelList);
+    Long getReviewCount(ExperienceModel experience);
+    List<Long> getListOfReviewCountByExperienceList(List<ExperienceModel> experienceModelList);
+    List<List<Long>> getListOfReviewCountByExperienceListAndCategoryId(List<List<ExperienceModel>> experienceModelList);
     List<Boolean> getListOfReviewHasImages(List<ReviewModel> reviewUserModelList);
     Page<ReviewModel> getReviewAndUser(ExperienceModel experience, Integer page);
     Optional<ReviewModel> getReviewById(Long reviewId);
